Add a kqueue-based implementation of IOLoop for BSD/Mac systems.
The main reason for this is to be able to detect when a connection has
been closed on the other side.
(http://github.com/facebook/tornado/issues/#issue/37)

This has been supported with epoll but not select, and it turned out to be
easier to add kqueue support than to figure out how to detect and handle
this case with select.

This change has not been tested in a production environment, but I have tested
it on my mac (10.6) and everything appears to work, and
IOStream.set_close_callback now works as expected.