type in options.define defaults to type of default if no type is given
type in tornado.options.define defaults to type of default, if no type is
given (type is str). This simplifies defines:
    >>> from tornado.options import define
    >>> define('port', 8888) # would be: define('port', 8888, int)
It seems more intuitive to me and reduces number of mistakes. This is also a
behaviour I would expect.
This only applies when default is not None, so there's no way a type would be
NoneType (unless explicitly given).