Stabilization of the period of ioloop.PeriodicCallback.
Previously, the timeout for the following event was computed by adding
callback_time to the actual time the callback finished, which caused
calls to the callback not being trully periodic.

This commit makes callback_time be a stable period by scheduling the
next timeout callback_time after the previous timeout was scheduled.