Adding Postgres support to database.py
database.py has been modified to have a MySQLConnection and PostgresConnection
class, creating connections for each database respectively. The Postgres
library models that of the MySql library with minor modifications where
necessary due to differences in the databases.

The differences are documented in the code, but two to note are that
Postgres by default no longer supports returning lastrowid, so
execute/executemany now returns None or the result of the query (if
RETURNING is specified).

Postgres also does not have max idle times by default so this was
accounted for in code.