Handle redirects before processing data and sending it to streaming_callback.
Previously, if you were using follow_redirects and streaming_callback
and hit a redirect with a non-empty body, that body would be passed
to streaming_callback as well.

As a side effect, this improves compatibility with some misconfigured
sites that send Content-Encoding: gzip on redirects without actually
compressing the body.