Made redirect_handler_class/error_handler_class configurable.
There are two new fields in Application.settings that are used in
tornado.web for the redirect/error handler classes. These default
to web.RedirectHandler and web.ErrorHandler. These are now handled
in the same was the static_handler_class is. This allows for
subclasses of Application and RequestHandler to customize this
logic more easily.