Add status_code to the OutputTransform.transform_first_chunk interface.
This is needed for correct support of the 304 status code, which
has no body and should not have either a Content-Length or
Transfer-Encoding.  This is a backwards-incompatible change to an
interface that was never technically private, but not included in the
documentation and as far as I can tell was never used outside
tornado itself.