Making Tornado compatible with local and `imp` importing
The internal module loading scheme is switched to relative module paths.
Since every module already has `absolute_import` this was not a big deal in terms of backward compatibility.

However, it is a big deal for allowing Tornado to be "embeded" into larger projects and distributed as such. Even the module folder can now be renamed to something like "mytornado" or "localtornado"

This also allows you to import tornado using `imp` module, allowing loading Tornado-containing code from locations outside of "." and sys.path.

In other words. This is awesome!