Supporting URL fragments.
Even though I'm fully aware that [RFC 3986](http://tools.ietf.org/html/rfc3986#section-3.5) specifically state that:
"...the fragment identifier is not used in the scheme-specific processing of a URI; instead, the fragment identifier is separated from the rest of the URI prior to a dereference, and thus the identifying information within the fragment itself is dereferenced solely by the user agent, regardless of the URI scheme."
It is somewhat dubious on whether a web framework like tornado should ignore the fragment. Users might need it to decide on what to serve. Consider this other part of the spec:
"As with any URI, use of a fragment identifier component does not imply that a retrieval action will take place.  A URI with a fragment identifier may be used to refer to the secondary resource without any implication that the primary resource is accessible or will ever be accessed."
All of this considered, this change only allows users to use fragments if they do see fit.