Update ca-certificates.crt to the current Mozilla version.
This file was generated by downloading
https://mxr.mozilla.org/mozilla/source/security/nss/lib/ckfw/builtins/certdata.txt?raw=1
in a web browser to avoid the chicken-and-egg problem with mk-ca-bundle's
built-in downloading, and then running "mk-ca-bundle.pl -n".

Note that we were not affected by the recently-fixed curl bug
(https://github.com/bagder/curl/commit/51f0b798fa), but only because
it had been so long since we had updated this file.