Reverse URL produces clean URLs even if path ends in /?
I often create url paths with /? at the end of them to make it easy for end users in case they type in / at the end of the url, but when I use reverse_url I do not want a /? to show up at the end. One option is to create two tornado.web.urls that both point to the same handler (one ends in /? and the named url doesn't), but I feel like this is a simple and clean addition and good for aesthetics. I usually override it in my projects, but others might benefit from it as well.