Update simple_httpclient.py
I have made a small change to simple_httpclient.py in method _on_close. Here an HTTPError is raised only if the connection object has an HTTPError in the stream. Otherwise, an Exception of the appropriate class is raised and the client will stop. I think we can write a small method to handle such exceptions in such a way that it prints the error message for the appropriate class without stopping the client program. On the other hand, I also feel that stopping the client program and displaying the entire stack trace would allow the programmer to debug better. Wdyt ?