Update websocket_test.py
The websocket on closing uses the simplehttpclient's _on_close method. Therefore, it was necessary that the tests for websocket_network_fail and websocket_network timeout be changed to expect an socket.error (or in case of an invalid address, socket.gaierror) and not a HTTPError. A new test has been added to check the condition where the host does not exist at the address mentioned raising a socket.gaierror with error number -5. This case has been added to verify the robustness of the patch.