Rename the new-in-4.0 gzip parameter to HTTPServer.
All the gzip-related parameters are now explicit about whether
they compress or decompress and whether they apply to requests or
responses.  For all the parameters that existed prior to 4.0
the old names are accepted as well, but for the new ones in 4.0
we don't need to worry about backwards-compatibility.

This is motivated by the potential confusion around the use of
gzip as a parameter to the Application constructor to indicate
compression of responses and to the HTTPServer constructor to
indicate decompression of requests.