Correctly handle 204 response codes that do not include a content-length.
If a server sends a 204 with no content-length we would wait for the server
to close the connection instead of ending the request (any such servers
are ignoring the "Connection: close" header we send, but apparently exist).

Move some content-length logic from simple_httpclient.py to http1connection.py.

Fix the client-side use of on_connection_close; this affects the
handling of any HTTPInputException.

This fixes regressions relative to Tornado 3.2.

Conflicts:
	tornado/websocket.py