Update KQueue.py
When have a big load of process attached to ioloop and it is running on Mac Os Yosemite (develop machine) when go to modify(self, fd, events) and it go to unregister(self, fd) it doesn't check if fd exist before pop from self._active.
I understand it should exist but doesn't and make internal issues and not register new fd.
Sometimes to exist and Python3 raise exception. Other time not exist but when go in except block now exist.