Tie the example for AsyncHTTPTestCase to real code.
Now all the reader has to do is copy one of the "Hello, world"
examples to hello.py, copy the test example to test_hello.py, import
AsyncHTTPTestCase, and run the test. It's still not entirely seamless
and idiot-proof, but it's progress!

Of note: the example test can now actually assert things about the
response it receives.