Add support for static extensionless files.
Provides support for serving files without their extension. The requested
URL /path/to/file will serve the file /path/to/file.html. Emulates the
behavior of Jekyll's Extensionless Permalinks:
http://jekyllrb.com/docs/permalinks/#extensionless-permalinks

The feature is off by default (when default_extension is set to None),
and is actived when default_extension is set to an extension (i.e. ".html").