Added SVG to compressible mime types
This adds the mime type `image/svg+xml` to the whitelist of compressible mime types for the tornado.web.GZipContentEncoding class.

SVG images are served with the mime type `image/svg+xml`, but are not compressed by the GZipContentEncoding transform. My first thought was to add all mime types ending with `+xml`, but that might need a bit more research.