locks: add priority support
wait() methods of Condition, Semaphore, BoundedSemaphore and Lock
classes now take priority value (0 by default). Waiters are woken
up in the order of priority (the smaller, the better), or, if
priorities are equal, in the order of addition. Thus, if priority
is not specified, the behavior of wait() is unchanged.