http1connection: add "Connection: close" header if appropriate
When HTTP1Connection acts as the server, it closes the connection
after writing the response if the client includes the
"Connection: close" header in the request, or if the `no_keep_alive`
option is set to True in the constructor.

According to https://tools.ietf.org/html/rfc7230#section-6.6

> The server SHOULD send a "close" connection option in its final
response on that connection.

It was possible for an Application to set the Connection header
appropriately. But it is very helpful for tornado to take care of
this automatically, particularly if "close" was specified in a
request header.