Correctly clear and create a new ioloop during autoreload (#2137)
* Correctly clear and create a new ioloop during autoreload

After the removal of the deprecated io_loop arguments from all functions,
the autoreload module stopped working as the wait function's usage was
missed in the refactor. This resulted in the start function receiving an
IOLoop object as its only argument which it then used as the check_time
argument resulting in errors further down the line when the check_time
is expected to be an int.

* Use the ioloop's add_callback function to start the ioloop on wait

There's no need to stop and clear the loop here, just add start as
a callback and start up the ioloop.