Exposes the currently running coroutine in gen.py
At the moment there is no easy way to provide per-instance context with
global scope if users decide to use coroutines in their
``tornado.web.RequestHandler`` methods.

Passing the request down the stack can be very verbose for large,
architecturally sound, code bases (service-oriented monoliths, for
example).

``tornado.stack_context.StackContext`` doesn't provide isolation between
requests when ``yield``s are encountered (and, hence, requests are
switched) over the course of a coroutine's lifetime.

This PR updates the currently running coroutine (and instance to which
it is potentially bound) to reflect which request is running in that
context, and generally what coroutine is running in others.

This is done by reference since a user can potentially make a copy of
that variable, leading to race conditions. That pitfall can be prevented
very simply in this way.