ioloop,asyncio: Be more careful to restore old asyncio loop
Any asyncio event loop that was created but neither closed nor set as
current for some thread will log a warning at the end of the test (in
a way that doesn't cause the test to fail). Fix a bug in
make_current() (in which multiple calls would clobber
self.old_asyncio) and use it more consistently when "current" status
changes.