Add server_hostname kwarg to `TCPClient.connect`
Allows overriding the `server_hostname` when connecting via TLS. This
can be useful when connecting to a server that hosts multiple SSL-based
services with distinct certificates, as the SNI (set by
`server_hostname` is used to determine which service to connect to.  If
unset, `server_hostname` still defaults to `host`.