Saving response
Saving the response allows a high level flow for api traces, for example if you want to make a collection of requests/ api / response. 
if your RequestHandler has several self.write due to error handling, its not clean to save in every self.write the response. Its better if self.write saves automatically
this output for a use with on_finish() method.