Add asyncSetUp and asyncTearDown to AsyncTestCase
These are coroutine equivalents of setUp and tearDown respectively, same
as found in unittest.IsolatedAsyncioTestCase in Python 3.8.

Use them to slightly simplify AsyncHTTPTestCase. Also add async_fetch to
the latter for use in @gen_test tests.

In addition to convenience, asyncSetUp/asyncTearDown help to avoid a
subtle issue: IOLoop.make_current() has no effect if there is a running
loop (asyncio.get_event_loop is hard-coded to return the running loop if
there is one). This makes setUp/tearDown run with a wrong current IOLoop.
This is not a problem for asyncSetUp/asyncTearDown, because the correct
loop is the running one. (Normally asyncio doesn't allow running multiple
loops, but this can be solved with nest-asyncio package.)