testing: Deprecate AsyncTestCase
The interfaces defined by this class rely on an event loop being
"current" even though it is not running; this is incompatible with
the future of the asyncio module.

It's too big a task to rewrite all the tests right now, so just
swallow deprecation warnings for "no current event loop".